Pass numeric values as doubles instead of integers
This approach is indeed very generic, but it’s more of a quick fix I
needed for my project.

+++ src/main/java/biz/paluch/logging/gelf/intern/GelfMessage.java
@@ -58,11 +58,11 @@ public class GelfMessage {
 
         for (Map.Entry<String, String> additionalField : additonalFields.entrySet()) {
             if (!ID_NAME.equals(additionalField.getKey())) {
-                // try adding the value as an integer
+                // try adding the value as a double
                 Object value;
                 try
                 {
-                    value = Integer.parseInt(additionalField.getValue());
+                    value = Double.parseDouble(additionalField.getValue());
                 }
                 catch (NumberFormatException ex)
                 {
